🚚 rename: Challenge 컴포넌트 이름 변경 (Challenges) #58
