+++ src/app/challenges/my_challenge/components/challenges.tsx
@@ -1,0 +1,53 @@
+import ChallengeCard from "./challengeCard";
+import { ChallengeStatesEnum } from "@/types/enums";
+import { IChallengeEnrolledList } from "@/types/challenge";
+
+export interface IChallengeInfo {
+  title: string;
+  participation: boolean;
+  participants: number;
+  startAt: Date;
+  endAt: Date;
+  fee: number;
+  achievement: number;
+}
+
+interface IChallengesProps {
+  challenges: IChallengeEnrolledList;
+  challengeState: ChallengeStatesEnum;
+}
+
+function Challenges({ challenges, challengeState }: IChallengesProps) {
+  if (challenges.inProgress.length === 0) {
+    return (
+      <span>참여 중인 챌린지가 없습니다. 새로운 챌린지에 참여해보세요!</span>
+    );
+  }
+
+  return (
+    <>
+      {challengeState == ChallengeStatesEnum.InProgress && (
+        <ChallengeCard
+          items={challenges.inProgress}
+          challengeState={challengeState}
+        />
+      )}
+
+      {challengeState == ChallengeStatesEnum.Scheduled && (
+        <ChallengeCard
+          items={challenges.scheduled}
+          challengeState={challengeState}
+        />
+      )}
+
+      {challengeState == ChallengeStatesEnum.Completed && (
+        <ChallengeCard
+          items={challenges.completed}
+          challengeState={challengeState}
+        />
+      )}
+    </>
+  );
+}
+
+export default Challenges;
